Fixed arguments count checks and conditional printing  in remove duplicated zkapp precondition row

+++ scripts/archive/zkapp_duplicated_precondition/remove_duplicated_zkapp_account_precondition_rows.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-if [ $# -gt 2 ]; then
+set -e
+
+if [ $# -ne 1 ]; then
     echo "Usage $0 archive-db"
     exit 0
 fi
@@ -33,7 +35,7 @@ echo "SELECT * FROM zkapp_account_update_body" | \
     psql -U postgres --csv -q -t "$ARCHIVE_DB" > $BODY_DATA_BCK
 
 
-echo "Creating temporary file of zkapp_account_preconditions table " "$PRECONDITION_DATA_TMP"
+echo "Creating temporary file of zkapp_account_preconditions table with duplicated rows" "$PRECONDITION_DATA_TMP"
 
 echo "SELECT  MIN(id) as id, $ALL_TABLES, MAX(id) as duplicated_id  \
 		FROM zkapp_account_precondition \
@@ -41,16 +43,23 @@ echo "SELECT  MIN(id) as id, $ALL_TABLES, MAX(id) as duplicated_id  \
 		HAVING COUNT(*) > 1" | \
     psql -U postgres --csv -q -t "$ARCHIVE_DB" > $PRECONDITION_DATA_TMP
 
-echo "Creating SQL scripts for removing duplications" "$PRECONDITION_UPDATE_SCRIPT" "and updating references" "$BODY_UPDATE_SCRIPT"
 
-cat $PRECONDITION_DATA_TMP | while IFS= read -r line 
-do  
-    ID_TO_REPLACE=$(echo "$line" | awk -F , '{print $1}');
-    ID_TO_REMOVE=$(echo "$line" | awk -F , '{print $10}');
-    echo -n .
-    echo "$ID_TO_REPLACE $ID_TO_REMOVE" | awk '{print "UPDATE zkapp_account_update_body SET zkapp_account_precondition_id = " $1 " WHERE zkapp_account_precondition_id =" $2 ";"}' >> $BODY_UPDATE_SCRIPT
-    echo "$ID_TO_REMOVE" | awk '{print "DELETE FROM zkapp_account_precondition WHERE id=" $1 ";"}' >> $PRECONDITION_UPDATE_SCRIPT
-done
+if [[ $(wc -l < $PRECONDITION_DATA_TMP) -gt 0 ]]; then
+    
+    echo "Creating SQL scripts for removing duplications" "$PRECONDITION_UPDATE_SCRIPT" "and updating references" "$BODY_UPDATE_SCRIPT"
 
-echo
-echo "SQL scripts are ready: $PRECONDITION_UPDATE_SCRIPT, $BODY_UPDATE_SCRIPT !"
+    cat $PRECONDITION_DATA_TMP | while IFS= read -r line 
+    do  
+        ID_TO_REPLACE=$(echo "$line" | awk -F , '{print $1}');
+        ID_TO_REMOVE=$(echo "$line" | awk -F , '{print $10}');
+        echo -n .
+        echo "$ID_TO_REPLACE $ID_TO_REMOVE" | awk '{print "UPDATE zkapp_account_update_body SET zkapp_account_precondition_id = " $1 " WHERE zkapp_account_precondition_id =" $2 ";"}' >> $BODY_UPDATE_SCRIPT
+        echo "$ID_TO_REMOVE" | awk '{print "DELETE FROM zkapp_account_precondition WHERE id=" $1 ";"}' >> $PRECONDITION_UPDATE_SCRIPT
+    done
+    
+    echo
+    echo "SQL scripts are ready: $PRECONDITION_UPDATE_SCRIPT, $BODY_UPDATE_SCRIPT !"
+else
+    echo
+    echo "No duplicated rows found. Database state is correct !"
+fi
